Updated Styling + Download All Tasks
Slightly Improving Styling + can now download JSON containing all tasks.

I put all the tasks in a String Stack, because the way pushing to a stack works (first element is the last element to be pushed). and i want the stack to be printed out in the actual proper order of tasks, and content of stacks